+++ source/duke3d/src/network.cpp
@@ -1596,7 +1596,7 @@ static void Net_SyncPlayer(ENetEvent *event)
     NET_75_CHECK++; // is it necessary to se event->peer->data to the new player index in Net_SyncPlayer?
     event->peer->data = (void *)(intptr_t)newPlayerIndex;
 
-    g_player[newPlayerIndex].netsynctime = totalclock;
+    g_player[newPlayerIndex].netsynctime = (int32_t) totalclock;
     g_player[newPlayerIndex].playerquitflag = 1;
 
     NET_75_CHECK++; // Need to think of something better when finding a remaining slot for players.
@@ -5146,8 +5146,8 @@ void Net_SendMessage(void)
         int32_t const text_x    = fullwidth >= (320 << 16) ? (320 << 16) - fullwidth : mpgametext_x;
         mpgametext(text_x, y, typebuf, 1, 2 | 8 | 16 | ROTATESPRITE_FULL16, 0, TEXT_YCENTER | TEXT_LITERALESCAPE);
         int32_t const cursor_x = text_x + width + textsc((tilesiz[SPINNINGNUKEICON].x << 14) + (1 << 16));
-        rotatesprite_fs(cursor_x, y, textsc(32768), 0, SPINNINGNUKEICON + ((totalclock >> 3) % 7),
-                        4 - (sintable[(totalclock << 4) & 2047] >> 11), 0, 2 | 8);
+        rotatesprite_fs(cursor_x, y, textsc(32768), 0, SPINNINGNUKEICON + (((int32_t) totalclock >> 3) % 7),
+                        4 - (sintable[((int32_t) totalclock << 4) & 2047] >> 11), 0, 2 | 8);
 
         if (hitstate == 1)
         {
